Turn some sqrt() into sqrtf() where double precision is not needed.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@1667 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/imdload.c
@@ -551,7 +551,7 @@ static BOOL _imd_load_points( char **ppFileData, iIMDShape *s )
 		s->radius = MAX(xmax, (MAX(ymax, zmax)));
 
 
-		s->sradius = (SDWORD)((float)sqrt( xmax*xmax + ymax*ymax + zmax*zmax));
+		s->sradius = (SDWORD)sqrtf( xmax*xmax + ymax*ymax + zmax*zmax);
 
 // START: tight bounding sphere
